+++ src/modules/platform-tools/hb-service/hb-service.service.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import { Injectable, NotFoundException, GoneException, BadRequestException } from '@nestjs/common';
+import * as stream from 'stream';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { ConfigService } from '../../../core/config/config.service';
 import { Logger } from '../../../core/logger/logger.service';
 
@@ -84,6 +85,38 @@ export class HbServiceService {
       this.logger.error(`Cannot download log file: Missing read permissions on "${this.configService.ui.log.path}".`);
       throw new BadRequestException('Cannot read log file. Check the log file permissions');
     }
-    return fs.createReadStream(this.configService.ui.log.path, 'utf8');
+
+    const removeColour = new stream.Transform({
+      transform(chunk, encoding, callback) {
+        callback(null, chunk.toString('utf8').replace(/\x1B\[([0-9]{1,3}(;[0-9]{1,2})?)?[mGK]/g, ''));
+      }
+    });
+
+    return fs.createReadStream(this.configService.ui.log.path, { encoding: 'utf8' })
+      .pipe(removeColour);
+  }
+
+  /**
+   * Truncate the log file
+   */
+  async truncateLogFile(username?: string) {
+    if (!await fs.pathExists(this.configService.ui.log.path)) {
+      this.logger.error(`Cannot truncate log file: "${this.configService.ui.log.path}" does not exist.`);
+      throw new BadRequestException(`Log file not found on disk.`);
+    }
+    try {
+      await fs.access(this.configService.ui.log.path, fs.constants.R_OK | fs.constants.W_OK);
+    } catch (e) {
+      this.logger.error(`Cannot truncate log file: Missing write permissions on "${this.configService.ui.log.path}".`);
+      throw new BadRequestException('Cannot access file. Check the log file permissions');
+    }
+
+    await fs.truncate(this.configService.ui.log.path);
+
+    setTimeout(() => {
+      this.logger.warn(`Homebridge log truncated by ${username || 'user'}.`);
+    }, 1000);
+
+    return { status: 0 };
   }
 }