+++ ui/src/app/modules/logs/logs.component.ts
@@ -10,8 +10,6 @@ import { ApiService } from '@/app/core/api.service';
 import { LogService } from '@/app/core/log.service';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ConfirmComponent } from '@/app/core/components/confirm/confirm.component';
-import { catchError } from 'rxjs/operators';
-import { throwError } from 'rxjs';
 
 @Component({
   selector: 'app-logs',
@@ -59,7 +57,7 @@ export class LogsComponent implements OnInit, OnDestroy {
     ref.componentInstance.confirmButtonLabel = this.$translate.instant('logs.label_download');
 
     ref.result.then(() => {
-      this.$api.get('/platform-tools/hb-service/download-log-file', { observe: 'response', responseType: 'blob' })
+      this.$api.get('/platform-tools/hb-service/log/download', { observe: 'response', responseType: 'blob' })
         .subscribe(
           (res: HttpResponse<any>) => {
             saveAs(res.body, 'homebridge.log.txt');
@@ -77,7 +75,31 @@ export class LogsComponent implements OnInit, OnDestroy {
     }).catch(() => {
       // do nothing
     });
+  }
 
+  truncateLogFile() {
+    const ref = this.$modal.open(ConfirmComponent);
+    ref.componentInstance.title = this.$translate.instant('logs.title_truncate_log_file');
+    ref.componentInstance.message = this.$translate.instant('logs.message_truncate_log_warning');
+    ref.componentInstance.confirmButtonLabel = this.$translate.instant('logs.label_truncate');
+
+    ref.result.then(() => {
+      this.$api.put('/platform-tools/hb-service/log/truncate', {})
+        .subscribe(
+          (res) => {
+            this.$toastr.success(
+              this.$translate.instant('logs.message_log_file_truncated'),
+              this.$translate.instant('toast.title_success'),
+            );
+            this.$log.term.clear();
+          },
+          (err: HttpErrorResponse) => {
+            this.$toastr.error(err.error.message || 'Failed to truncate log file', this.$translate.instant('toast.title_error'));
+          },
+        );
+    }).catch(() => {
+      // do nothing
+    });
   }
 
 }